replace ImageUpload with NoSSR version
