+++ components/Org/OrgDetailForm.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 import { FormattedMessage } from 'react-intl'
 import RichTextEditor from '../Form/Input/RichTextEditor'
-import ImageUpload from '../Upload/ImageUploadNoSSR'
+import ImageUpload from '../Upload/ImageUpload'
 import PageTitle from '../../components/LandingPageComponents/PageTitle.js'
 import { domainRegex } from '../../lib/fieldValidation'
 import {
