--- components/Upload/ImageUploadNoSSR.js
@@ -1,7 +1,0 @@
-import dynamic from 'next/dynamic'
-
-const ImageUpload = dynamic(() => import('./ImageUpload'), {
-  ssr: false
-})
-
-export default () => <ImageUpload />
